+++ lib/blocks.js
@@ -16,16 +16,7 @@ var C_OPEN_BRACKET = 91;
 
 var InlineParser = require('./inlines');
 
-var reHtmlBlockOpen = [
-   /./, // dummy for 0
-   /^<(?:script|pre|style)(?:\s|>|$)/i,
-   /^<!--/,
-   /^<[?]/,
-   /^<![A-Z]/,
-   /^<!\[CDATA\[/,
-   /^<[/]?(?:address|article|aside|base|basefont|blockquote|body|caption|center|col|colgroup|dd|details|dialog|dir|div|dl|dt|fieldset|figcaption|figure|footer|form|frame|frameset|h1|head|header|hr|html|iframe|legend|li|link|main|menu|menuitem|meta|nav|noframes|ol|optgroup|option|p|param|section|source|title|summary|table|tbody|td|tfoot|th|thead|title|tr|track|ul)(?:\s|[/]?[>]|$)/i,
-    new RegExp('^(?:' + OPENTAG + '|' + CLOSETAG + ')\s*$', 'i')
-];
+var reHtmlBlockOpen = require('./common').reHtmlBlockOpen;
 
 var reHtmlBlockClose = [
    /./, // dummy for 0
