+++ lib/common.js
@@ -27,6 +27,17 @@ var HTMLTAG = "(?:" + OPENTAG + "|" + CLOSETAG + "|" + HTMLCOMMENT + "|" +
         PROCESSINGINSTRUCTION + "|" + DECLARATION + "|" + CDATA + ")";
 var reHtmlTag = new RegExp('^' + HTMLTAG, 'i');
 
+var reHtmlBlockOpen = [
+   /./, // dummy for 0
+   /^<(?:script|pre|style)(?:\s|>|$)/i,
+   /^<!--/,
+   /^<[?]/,
+   /^<![A-Z]/,
+   /^<!\[CDATA\[/,
+   /^<[/]?(?:address|article|aside|base|basefont|blockquote|body|caption|center|col|colgroup|dd|details|dialog|dir|div|dl|dt|fieldset|figcaption|figure|footer|form|frame|frameset|h1|head|header|hr|html|iframe|legend|li|link|main|menu|menuitem|meta|nav|noframes|ol|optgroup|option|p|param|section|source|title|summary|table|tbody|td|tfoot|th|thead|title|tr|track|ul)(?:\s|[/]?[>]|$)/i,
+    new RegExp('^(?:' + OPENTAG + '|' + CLOSETAG + ')\s*$', 'i')
+];
+
 var reBackslashOrAmp = /[\\&]/;
 
 var ESCAPABLE = '[!"#$%&\'()*+,./:;<=>?@[\\\\\\]^_`{|}~-]';
@@ -96,6 +107,7 @@ module.exports = { unescapeString: unescapeString,
                    normalizeURI: normalizeURI,
                    escapeXml: escapeXml,
                    reHtmlTag: reHtmlTag,
+                   reHtmlBlockOpen: reHtmlBlockOpen,
                    OPENTAG: OPENTAG,
                    CLOSETAG: CLOSETAG,
                    ENTITY: ENTITY,
