+++ lib/inlines.js
@@ -10,6 +10,8 @@ var fromCodePoint = require('./from-code-point.js');
 var decodeHTML = require('entities').decodeHTML;
 require('string.prototype.repeat'); // Polyfill for String.prototype.repeat
 
+var reHtmlBlockOpen = require('./common').reHtmlBlockOpen;
+
 // Constants for character codes:
 
 var C_NEWLINE = 10;
@@ -213,6 +215,15 @@ var parseHtmlTag = function(block) {
         var node = new Node('html_inline');
         node._literal = m;
         block.appendChild(node);
+
+        var blockType;
+        for (blockType = 1; blockType <= 7; blockType++) {
+            if(reHtmlBlockOpen[blockType].test(m) && blockType < 7) {
+              node._htmlBlockType = blockType;
+              break;
+            }
+        }
+
         return true;
     }
 };
